softfloat: bf16: change parameter type to match function name
Though we use float16_t for vs2 in vfncvtbf16_f_f_w.h and vfncvtbf16_sat_f_f_w.h,
 it can pass compilation since they are the same type

+++ softfloat/bf16_to_e4m3.c
@@ -41,7 +41,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "specialize.h"
 #include "softfloat.h"
 
-e4m3_t bf16_to_e4m3( float16_t a, bool saturationMode )
+e4m3_t bf16_to_e4m3( bfloat16_t a, bool saturationMode )
 {
    return f32_to_e4m3( bf16_to_f32( a ), saturationMode );
 }
